refactor(frontend): rename ccn 0044 files
