+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ccn/0044/coefficients.js
