+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/components/IndemniteConventionnelle.js
@@ -49,7 +49,7 @@ function IndemniteCCn({
       )}
       <br />
       <br />
-      <Button> Recommencer une simulation </Button>
+      <Button>Recommencer une simulation</Button>
     </Container>
   );
 }
